Scripts/TrainCodecEncoder/requirements.txt
torch>=2.1.0
torchaudio>=2.1.0
safetensors>=0.4.0
openai-whisper>=20231117
soundfile>=0.12.0
librosa>=0.10.0
pesq>=0.0.4
pystoi>=0.4.0
tqdm>=4.65.0
wandb>=0.16.0
